requirements.txt
selenium==4.21.0
pytest
pytest-html
python-dotenv
